// src/app/models/DTO/RegUserDto.ts
export class RegUserDto {
    id: number;
    name: string;
    firstName: string;
    middleName: string;
    lastName: string;
    email: string;
    identityNum: string;
    phoneNumber: string;
    userName: string;
    password: string;
    age: number;
    address: string;
    cityId: number;
    cntryId: number;
    birthDay: Date;
    partyId: number
    permsIdList: number[]
    repId: number;
    typeId: number;
    organisationName: string;
    organisationAddress: string;
    commercialRegNum: number;
    repName: string;
    idNo: string;
}